+++ lib/mdqt/client.rb
@@ -21,9 +21,10 @@ module MDQT
 
       @base_url   = base_url
       @verbose    = options[:verbose] || false
+      @explain    = options[:explain] || false
       @cache_type = options[:cache_type] || :none
 
-      @md_service = MetadataService.new(@base_url, verbose: @verbose, cache_type: @cache_type)
+      @md_service = MetadataService.new(@base_url, verbose: @verbose, cache_type: @cache_type, explain: @explain)
 
     end
 
@@ -39,6 +40,10 @@ module MDQT
 
     def verbose?
       @verbose
+    end
+
+    def explain?
+      @explain
     end
 
     def cache_type