+++ lib/mdqt/client/metadata_validator.rb
@@ -3,8 +3,6 @@ module MDQT
 
     class MetadataValidator
 
-      require 'xmldsig'
-
       def initialize(options = {})
         @certs = options[:certs] || []
       end