+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Invalid SHA1 transformed identitifiers can't be sent
 - 500 errors at the server will be shown correctly
 
+### Fixes
+- Don't show empty identifier in OK message after downloading aggregate
+
 ## 0.2.1
 
 ### Fixes