+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.0
 
 ### Improvements
+- New signature verification, using `--verify-with` option for get command
 - New `reset` command to clear all cached metadata
 - Server's 304 responses for cached files are handled correctly
 - Invalid SHA1 transformed identitifiers can't be sent