+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 - New signature verification, using `--verify-with` option for get command
 - New `reset` command to clear all cached metadata
+- Coloured feedback
 - Server's 304 responses for cached files are handled correctly
 - Invalid SHA1 transformed identitifiers can't be sent
 - 500 errors at the server will be shown correctly